refactor: updated admin file and function names
