+++ src/components/pages/admin-mode/area-info-dialog/show-area-info-dialog.module.ts
