+++ src/components/pages/admin-mode/person-info-dialog/show-person-info-dialog.module.ts
